% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon.cds.extraction.R
\name{exon.cds.extraction}
\alias{exon.cds.extraction}
\title{exon.cds.extraction}
\usage{
exon.cds.extraction(input)
}
\arguments{
\item{input}{annotation file outputted by snpEff.}
}
\description{
Extracting exon or cds informtaion of primary annotation from snpEff.
}
\examples{

exon.cds.extraction(input = "mid.anno")

exon.cds.extraction(input = "__breakpoint.annotated")

}
