% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq2png.R
\name{fq2png}
\alias{fq2png}
\title{fq2png}
\usage{
fq2png(seq, aln, nam)
}
\arguments{
\item{seq}{raw fastq suquence of breakpoint}

\item{aln}{Transformed cigar records of breakpoint reads}

\item{nam}{prefixed name of output file}
}
\description{
Converted fastq sequnce into png file.
}
\examples{

fq2png(seq = "example.EML4_intron6---ALK_exon20.txt", aln = "breakpoint.reads"
,nam = "example.EML4_intron6---ALK_exon20")

}
